requirements.txt
python-telegram-bot==21.*
gspread
google-auth
